refactor(decompose): split webapp/e2e-tests/electron/for_feature_development_not_LT_verification/terminals/stop-gate/electron-stop-gate-hooks.spec.ts into folder
